accuknox_main/requirements.txt
django==4.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1